refactor(presentation): inline CliAgentController logic into agent-cli-presenter'
